Add material table complete

+++ src/components/RowActions.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import PropTypes from "prop-types";
 import TableCell from "@mui/material/TableCell";
 import DeleteIcon from "@mui/icons-material/Delete";
@@ -6,16 +6,16 @@ import EditIcon from "@mui/icons-material/Edit";
 import IconButton from "@mui/material/IconButton";
 import Stack from "@mui/material/Stack";
 
+import AlertDialog from "./AlertDialog";
+
 const RowActions = ({ onEdit, onDelete }) => {
+  const [showAlertDialog, setShowAlertDialog] = useState(false);
   return (
     <TableCell align="left" padding="none">
       <Stack direction={"row"}>
         {onEdit && (
           <IconButton
-            onClick={(e) => {
-              e.stopPropagation();
-              onEdit();
-            }}
+            onClick={onEdit}
             sx={{ color: "#3582d6a8", fontSize: "16px" }}
           >
             <EditIcon />
@@ -23,16 +23,22 @@ const RowActions = ({ onEdit, onDelete }) => {
         )}
         {onDelete && (
           <IconButton
-            onClick={(e) => {
-              e.stopPropagation();
-              onDelete();
-            }}
+            onClick={() => setShowAlertDialog(true)}
             sx={{ color: "#d54565a8", fontSize: "16px" }}
           >
             <DeleteIcon />
           </IconButton>
         )}
       </Stack>
+      <AlertDialog
+        open={showAlertDialog}
+        message={"Are you sure you wanted to delete records"}
+        handleDialogClose={() => setShowAlertDialog(false)}
+        onPositiveResponse={() => {
+          onDelete();
+          setShowAlertDialog(false);
+        }}
+      />
     </TableCell>
   );
 };
